refactor: Remove unused components and update examples
- Examples now use UDFs for specific operations rather than unlikely to
  be useful components.
- New example demonstrates how to use the Python SDK as a library to
  create components with business logic.
- Moved examples from `stepflow-rs/examples` to `examples` since they
  are general examples for using stepflow.
