+++ examples/basic/input.json
@@ -1,0 +1,1 @@
+{"m": 3, "n": 4}
