+++ examples/basic/stepflow-config.yml
@@ -4,4 +4,4 @@ plugins:
   - name: python
     type: stdio
     command: uv
-    args: ["--project", "../../../sdks/python", "run", "stepflow_sdk"]
+    args: ["--project", "../../sdks/python", "run", "stepflow_sdk"]
