+++ examples/custom-python-components/stepflow-config.yml
@@ -1,0 +1,7 @@
+plugins:
+  - name: builtin
+    type: builtin
+  - name: custom
+    type: stdio
+    command: uv
+    args: ["--project", "../../sdks/python", "run", "python", "custom_server.py"]
