+++ examples/data-pipeline/stepflow-config.yml
@@ -2,6 +2,6 @@ plugins:
   - name: python
     type: stdio
     command: uv
-    args: ["--project", "../../../sdks/python", "run", "stepflow_sdk"]
+    args: ["--project", "../../sdks/python", "run", "stepflow_sdk"]
   - name: builtin
     type: builtin
