+++ examples/eval/stepflow-config.yml
@@ -1,0 +1,7 @@
+plugins:
+  - name: builtin
+    type: builtin
+  - name: python
+    type: stdio
+    command: uv
+    args: ["--project", "../../sdks/python", "run", "stepflow_sdk"]
