+++ examples/file-loading/load-pipeline.yaml
@@ -1,0 +1,99 @@
+name: "File Loading Sales Pipeline"
+description: "Load sales data from file and process it"
+
+input_schema:
+  type: object
+  properties:
+    data_file_path:
+      type: string
+      description: "Path to the sales data file"
+    target_revenue:
+      type: number
+      description: "Target revenue for analysis"
+
+steps:
+  # Load the sales data from file
+  - id: load_sales_data
+    component: "builtin://load_file"
+    input:
+      path: { $from: "$input", path: "data_file_path" }
+
+  # Extract the sales_data array using UDF
+  - id: extract_sales_udf
+    component: "builtin://put_blob"
+    input:
+      data:
+        $literal:
+          code: "input['file_data']['sales_data']"
+
+  - id: extract_sales_array
+    component: python://udf
+    input:
+      blob_id: { $from: { step: extract_sales_udf }, path: "blob_id" }
+      input:
+        file_data: { $from: { step: "load_sales_data" }, path: "data" }
+
+  # Now process the sales data array
+  - id: sum_revenue_udf
+    component: "builtin://put_blob"
+    input:
+      data:
+        $literal:
+          code: "sum(item['revenue'] for item in input['sales_data'])"
+
+  - id: calculate_total_revenue
+    component: python://udf
+    input:
+      blob_id: { $from: { step: sum_revenue_udf }, path: "blob_id" }
+      input:
+        sales_data: { $from: { step: "extract_sales_array" }, path: "result" }
+
+  - id: count_sales_udf
+    component: "builtin://put_blob"
+    input:
+      data:
+        $literal:
+          code: "len(input['sales_data'])"
+
+  - id: count_sales
+    component: python://udf
+    input:
+      blob_id: { $from: { step: count_sales_udf }, path: "blob_id" }
+      input:
+        sales_data: { $from: { step: "extract_sales_array" }, path: "result" }
+
+  - id: average_sale_udf
+    component: "builtin://put_blob"
+    input:
+      data:
+        $literal:
+          code: "sum(item['revenue'] for item in input['sales_data']) / len(input['sales_data']) if input['sales_data'] else 0"
+
+  - id: calculate_average_sale
+    component: python://udf
+    input:
+      blob_id: { $from: { step: average_sale_udf }, path: "blob_id" }
+      input:
+        sales_data: { $from: { step: "extract_sales_array" }, path: "result" }
+
+  - id: divide_udf
+    component: "builtin://put_blob"
+    input:
+      data:
+        $literal:
+          code: "input['a'] / input['b'] if input['b'] != 0 else 0"
+
+  - id: calculate_performance_ratio
+    component: python://udf
+    input:
+      blob_id: { $from: { step: divide_udf }, path: "blob_id" }
+      input:
+        a: { $from: { step: "calculate_total_revenue" }, path: "result" }
+        b: { $from: { workflow: input }, path: "target_revenue" }
+
+output:
+  file_metadata: { $from: "load_sales_data", path: "metadata" }
+  total_revenue: { $from: "calculate_total_revenue", path: "result" }
+  sales_count: { $from: "count_sales", path: "result" }
+  average_sale: { $from: "calculate_average_sale", path: "result" }
+  performance_ratio: { $from: "calculate_performance_ratio", path: "result" }
