+++ examples/openai/openai_chat.yaml
