+++ stepflow-rs/crates/stepflow-main/tests/commands/test_submit.rs
@@ -26,7 +26,7 @@ fn test_submit_with_file_input() {
             .arg("submit")
             .arg("--url=http://localhost:9999")
             .arg("--flow=tests/mock/basic.yaml")
-            .arg("--input=examples/eval/input.json")
+            .arg("--input=../../../../examples/eval/input.json")
     );
 }
