--- stepflow-rs/examples/eval/math-eval.yaml
@@ -1,42 +1,0 @@
-name: "Math Eval Test"
-description: "Test eval with actual computation"
-
-input_schema:
-  type: object
-  properties:
-    numbers:
-      type: array
-
-steps:
-  - id: nested_math
-    component: "builtin://eval"
-    input:
-      workflow:
-        $literal:
-          name: "Nested Math Workflow"
-          steps:
-            - id: compute_sum
-              component: python://add
-              input:
-                a: { $from: { workflow: input }, path: "x" }
-                b: { $from: { workflow: input }, path: "y" }
-          output:
-            sum_result: { $from: { step: compute_sum }, path: "result" }
-      input:
-        x: 5
-        y: 3
-
-output:
-  math_result: { $from: { step: nested_math }, path: "result" }
-
-test:
-  cases:
-  - name: nested_math_calculation
-    description: Test nested workflow evaluation with math
-    input:
-      numbers: [5, 3]
-    output:
-      outcome: success
-      result:
-        math_result:
-          sum_result: 8
