--- stepflow-rs/examples/file-loading/load-pipeline.yaml
@@ -1,56 +1,0 @@
-name: "File Loading Sales Pipeline"
-description: "Load sales data from file and process it"
-
-input_schema:
-  type: object
-  properties:
-    data_file_path:
-      type: string
-      description: "Path to the sales data file"
-    target_revenue:
-      type: number
-      description: "Target revenue for analysis"
-
-steps:
-  # Load the sales data from file
-  - id: load_sales_data
-    component: "builtin://load_file"
-    input:
-      path: { $from: "$input", path: "data_file_path" }
-
-  # Extract the sales_data array using Python
-  - id: extract_sales_array
-    component: python://extract_sales_data
-    input:
-      input: { $from: "load_sales_data", path: "data" }
-
-  # Now process the sales data array
-  - id: calculate_total_revenue
-    component: python://sum_field
-    input:
-      input: { $from: "extract_sales_array", path: "input" }
-      field: "revenue"
-
-  - id: count_sales
-    component: python://count_items
-    input:
-      input: { $from: "extract_sales_array", path: "input" }
-
-  - id: calculate_average_sale
-    component: python://average_field
-    input:
-      input: { $from: "extract_sales_array", path: "input" }
-      field: "revenue"
-
-  - id: calculate_performance_ratio
-    component: python://divide
-    input:
-      a: { $from: "calculate_total_revenue", path: "result" }
-      b: { $from: "$input", path: "target_revenue" }
-
-output:
-  file_metadata: { $from: "load_sales_data", path: "metadata" }
-  total_revenue: { $from: "calculate_total_revenue", path: "result" }
-  sales_count: { $from: "count_sales", path: "result" }
-  average_sale: { $from: "calculate_average_sale", path: "result" }
-  performance_ratio: { $from: "calculate_performance_ratio", path: "result" }
