--- stepflow-rs/examples/python/stepflow-config.yml
@@ -1,5 +1,0 @@
-plugins:
-  - name: python
-    type: stdio
-    command: uv
-    args: ["--project", "../../../sdks/python", "run", "stepflow_sdk"]
